stop using two codepaths for computing file and glob signatures, let the backend always deal with it, so it's impossible to get difference in implementation between frontend and backend.
