+++ src/Driver.cpp
@@ -18,6 +18,7 @@
 #include "Hash.hpp"
 #include "Profiler.hpp"
 #include "NodeResultPrinting.hpp"
+#include "FileSign.hpp"
 
 #include <time.h>
 #include <stdio.h>
@@ -260,84 +261,12 @@ static bool DriverCheckDagSignatures(Driver* self)
     }
   }
 
-  // Helper for directory iteration + memory allocation of strings.  We need to
-  // buffer the filenames as we need them in sorted order to ensure the results
-  // are consistent between runs.
-  struct IterContext
-  {
-    MemAllocLinear       *m_Allocator;
-    MemAllocHeap         *m_Heap;
-    Buffer<const char *>  m_Dirs;
-    Buffer<const char *>  m_Files;
-
-    void Init(MemAllocHeap* heap, MemAllocLinear* linear)
-    {
-      m_Allocator = linear;
-      m_Heap      = heap;
-      BufferInit(&m_Dirs);
-      BufferInit(&m_Files);
-    }
-
-    void Destroy()
-    {
-      BufferDestroy(&m_Files, m_Heap);
-      BufferDestroy(&m_Dirs, m_Heap);
-    }
-
-    static void Callback(void* user_data, const FileInfo& info, const char* path)
-    {
-      IterContext* self = (IterContext*) user_data;
-      char* data = StrDup(self->m_Allocator, path);
-      Buffer<const char*>* target = info.IsDirectory() ? &self->m_Dirs : &self->m_Files;
-      BufferAppendOne(target, self->m_Heap, data);
-    }
-
-    static int SortStringPtrs(const void* l, const void* r)
-    {
-      return strcmp(*(const char**)l, *(const char**)r);
-    }
-  };
-
   // Check directory listing fingerprints
   // Note that the digest computation in here must match the one in LuaListDirectory
   // The digests computed there are stored in the signature block by Lua code.
   for (const DagGlobSignature& sig : dag_data->m_GlobSignatures)
   {
-    const char* path = sig.m_Path;
-
-    // Set up to rewind allocator for each loop iteration
-    MemAllocLinearScope mem_scope(&self->m_Allocator);
-
-    // Set up context
-    IterContext ctx;
-    ctx.Init(&self->m_Heap, &self->m_Allocator);
-
-    // Get directory data
-    ListDirectory(path, &ctx, IterContext::Callback);
-
-    // Sort data
-    qsort(ctx.m_Dirs.m_Storage, ctx.m_Dirs.m_Size, sizeof(const char*), IterContext::SortStringPtrs);
-    qsort(ctx.m_Files.m_Storage, ctx.m_Files.m_Size, sizeof(const char*), IterContext::SortStringPtrs);
-
-    // Compute digest
-    HashState h;
-    HashInit(&h);
-    for (const char* p : ctx.m_Dirs)
-    {
-      HashAddPath(&h, p);
-      HashAddSeparator(&h);
-    }
-
-    for (const char* p : ctx.m_Files)
-    {
-      HashAddPath(&h, p);
-      HashAddSeparator(&h);
-    }
-
-    HashDigest digest;
-    HashFinalize(&h, &digest);
-
-    ctx.Destroy();
+    HashDigest digest = CalculateGlobSignatureFor(sig.m_Path, &self->m_Heap, &self->m_Allocator);
 
     // Compare digest with the one stored in the signature block
     if (0 != memcmp(&digest, &sig.m_Digest, sizeof digest))
@@ -1181,6 +1110,7 @@ bool DriverSaveBuildState(Driver* self)
   BinarySegmentWriteInt32(main_seg, entry_count);
   BinarySegmentWritePointer(main_seg, guid_ptr);
   BinarySegmentWritePointer(main_seg, state_ptr);
+  BinarySegmentWriteUint32(main_seg, StateData::MagicNumber);
 
   // Unmap old state data.
   MmapFileUnmap(&self->m_StateFile);
